% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reporting_units.R
\name{get_reporting_units}
\alias{get_reporting_units}
\title{get_reporting_units}
\usage{
get_reporting_units(ru_url)
}
\arguments{
\item{ru_url}{character, a url that will return reporting unit data. 
Will work with a path to a local file too.}
}
\value{
\code{data.frame} containing wade data
}
\description{
Get list of reporting units from wade service.
}
\examples{
ru_url <- 'http://www.westernstateswater.org/Wyoming/WADE/v0.2/GetCatalog/GetCatalog_GetAll.php?orgid=WYWDC'
get_reporting_units(ru_url)

}
