% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wade_data.R
\name{get_wade_data}
\alias{get_wade_data}
\title{get_wade_data}
\usage{
get_wade_data(wade_url)
}
\arguments{
\item{wade_url}{character, a url that will return wade data. 
Will work with a path to a local file too.}
}
\value{
\code{data.frame} containing wade data
}
\description{
Get data from a wade service.
}
\examples{
wade_url <- paste0("https://water.utah.gov/DWRE/WADE/v0.2/GetSummary/GetSummary.php?",
                   "loctype=REPORTUNIT&loctxt=00-02-02&orgid=utwre",
                   "&reportid=2005_ConsumptiveUse&datatype=ALL")
get_wade_data(wade_url)

}
